\name{Datevec}
\alias{Datevec}
\title{
Convert serial date number to date, same as Matlab
}
\description{
\code{Datevec} Convert serial date number to date 
with the same values as Matlab's datevec
}
\usage{
Datevec(s)
}
\arguments{
  \item{s}{ serial date number}
}
\value{date object
}
\examples{
Datevec(724789)
}
\author{
Jose Gama
}
\keyword{ programming }
