\name{DatevecV}
\alias{DatevecV}
\title{
Convert serial date number to list(year,month,day), same as Matlab
}
\description{
\code{DatevecV} Convert serial date number to 
list(year,month,day) with the same values as Matlab's datevec
}
\usage{
DatevecV(s)
}
\arguments{
  \item{s}{ list(year,month,day)}
}
\value{date object
}
\examples{
DatevecV(724789)
}
\author{
Jose Gama
}
\keyword{ programming }
