\name{Earthtide}
\alias{Earthtide}
\title{
Predict displacements from solid earth tides
}
\description{
\code{Earthtide} predicts displacements from solid earth tides.
}
\usage{
Earthtide(latIn, lonIn, timeSeq, boolHorizontalMotion = FALSE)
}
\arguments{
  \item{latIn}{latitude}
  \item{lonIn}{longitude}
  \item{timeSeq}{time in UTC time units}
  \item{boolHorizontalMotion}{optional, TRUE=return a list with vertical, Noth and East displacements}
}
\source{
Chad Greene, 2015
Solid Earth Tide Predictions
\url{http://www.mathworks.com/matlabcentral/fileexchange/50327-sold-earth-tide-predictions}
}
\references{
Chad Greene, 2015
Solid Earth Tide Predictions
\url{http://www.mathworks.com/matlabcentral/fileexchange/50327-sold-earth-tide-predictions}
}
\author{
Jose Gama
}
\keyword{ programming }
