\name{GetLeapDates}
\alias{GetLeapDates}
\title{
Leap dates from the UTC to GPS Time Conversion Table
}
\description{
\code{GetLeapDates} return the leap dates from the 
UTC to GPS Time Conversion Table, fixing the first value
}
\usage{
GetLeapDates()
}
\value{leap dates
}
\author{
Jose Gama
}
\keyword{ programming }
