\name{LongTerm}
\alias{LongTerm}
\title{Long Term Tidal Constituent Table}
\description{\code{LongTerm} Long Term Tidal Constituent Data}
\usage{data(LongTerm)}
\format{
  A data frame with 5 values on the following 5 variables.
  \describe{ 	 	 	
    \item{Tidal.constituent}{Tidal constituent}
    \item{Period}{Period}
    \item{Period.Unit}{Period unit}
    \item{Vertical.amplitude..mm.}{Vertical amplitude in mm}
    \item{Horizontal.amplitude..mm.}{Horizontal amplitude in mm}
  }
}
\source{
Earth tide - Wikipedia, the free encyclopedia
\url{https://en.wikipedia.org/wiki/Earth_tide}
}
\references{
Earth tide - Wikipedia, the free encyclopedia
\url{https://en.wikipedia.org/wiki/Earth_tide}
}
\author{Jose Gama}
\keyword{datasets}
