\name{SemiDiurnal}
\alias{SemiDiurnal}
\title{Semi Diurnal Tidal Constituent Table}
\description{\code{SemiDiurnal} Semi Diurnal Tidal Constituent Data}
\usage{data(SemiDiurnal)}
\format{
  A data frame with 4 values on the following 4 variables.
  \describe{ 	 	 	
    \item{Tidal.constituent}{Tidal constituent}
    \item{Period..hr.}{Period in hours}
    \item{Vertical.amplitude..mm.}{Vertical amplitude in mm}
    \item{Horizontal.amplitude..mm.}{Horizontal amplitude in mm}
  }
}
\source{
Earth tide - Wikipedia, the free encyclopedia
\url{https://en.wikipedia.org/wiki/Earth_tide}
}
\references{
Earth tide - Wikipedia, the free encyclopedia
\url{https://en.wikipedia.org/wiki/Earth_tide}
}
\author{Jose Gama}
\keyword{datasets}
