\name{UTC2GPS}
\alias{UTC2GPS}
\title{
Convert UTC(GMT) time tags to GPS time accounting for leap seconds
}
\description{
\code{UTC2GPS} converts UTC(GMT) time tags to GPS time accounting for leap seconds.
}
\usage{
UTC2GPS(utctime, asMATLAB=FALSE)
}
\arguments{
  \item{utctime}{time in UTC time units}
  \item{asMATLAB}{optional boolean, TRUE=output is MATLAB time units}
}
\source{
Ian M. Howat, 2008
utc2gps
\url{http://www.mathworks.com/matlabcentral/fileexchange/21193-utc-to-gps-time-coverter/content/utc2gps.m}
}
\references{
Ian M. Howat, 2008
utc2gps
\url{http://www.mathworks.com/matlabcentral/fileexchange/21193-utc-to-gps-time-coverter/content/utc2gps.m}
}
\author{
Jose Gama
}
\keyword{ programming }
