\name{UTCtoGPSTimeConversion}
\alias{UTCtoGPSTimeConversion}
\title{UTC to GPS Time Conversion Table}
\description{\code{UTCtoGPSTimeConversion} UTC to GPS Time Conversion Table}
\usage{data(UTCtoGPSTimeConversion)}
\format{
  A data frame with 71329 observations on the following 16 variables.
  \describe{
    \item{yearLimitFrom}{Limits of validity (at 0h UTC) from}
    \item{yearLimitTo}{Limits of validity (at 0h UTC) to}
    \item{TAIminusUTC}{TAI - UTC (s)}
    \item{GPSminusUTC}{GPS - UTC (s)}
  }
}
\source{
confluence, 2015
UTC to GPS Time Correction
\url{https://confluence.qps.nl/display/KBE/UTC+to+GPS+Time+Correction}
}
\references{
confluence, 2015
UTC to GPS Time Correction
\url{https://confluence.qps.nl/display/KBE/UTC+to+GPS+Time+Correction}
}
\author{Jose Gama}
\keyword{datasets}
